[#57] 그룹 관리 API 추가
- 오늘의 추천 그룹 목록 조회 수정
