+++ src/main/java/io/driver/codrive/modules/room/domain/RoomRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoomRepository extends JpaRepository<Room, Long> {
+public interface RoomRepository extends JpaRepository<Room, Long>, RoomRepositoryCustom {
 	Page<Room> findAll(Pageable pageable);
 	Optional<Room> findByUuid(String uuid);
 	Page<Room> findByTitleContaining(String keyword, Pageable pageable);
