+++ src/main/java/io/driver/codrive/modules/room/domain/RoomRepositoryCustom.java
@@ -1,0 +1,7 @@
+package io.driver.codrive.modules.room.domain;
+
+import java.util.List;
+
+public interface RoomRepositoryCustom {
+	List<Room> getRoomsByLanguageExcludingOwnRoom(Long languageId, Long userId);
+}
